Remove static cast

+++ cpp/dolfinx/mesh/EntityMap.h
@@ -48,8 +48,7 @@ public:
           "No index map for entities, call `Topology::create_entities("
           + std::to_string(_dim) + ")");
     }
-    std::size_t num_ents
-        = static_cast<std::size_t>(e_map->size_local() + e_map->num_ghosts());
+    std::size_t num_ents = e_map->size_local() + e_map->num_ghosts();
     if (num_ents != _sub_topology_to_topology.size())
       throw std::runtime_error(
           "Size mismatch between `sub_topology_to_topology` and index map.");
